Extract ParserBase from the Parser
