+++ h2/src/main/org/h2/command/Parser.java
@@ -39,7 +39,6 @@ import static org.h2.command.Token.SMALLER;
 import static org.h2.command.Token.SMALLER_EQUAL;
 import static org.h2.command.Token.SPATIAL_INTERSECTS;
 import static org.h2.command.Token.TILDE;
-import static org.h2.command.Token.TOKENS;
 import static org.h2.util.ParserUtil.ALL;
 import static org.h2.util.ParserUtil.AND;
 import static org.h2.util.ParserUtil.ANY;
@@ -70,7 +69,6 @@ import static org.h2.util.ParserUtil.EXCEPT;
 import static org.h2.util.ParserUtil.EXISTS;
 import static org.h2.util.ParserUtil.FALSE;
 import static org.h2.util.ParserUtil.FETCH;
-import static org.h2.util.ParserUtil.FIRST_KEYWORD;
 import static org.h2.util.ParserUtil.FOR;
 import static org.h2.util.ParserUtil.FOREIGN;
 import static org.h2.util.ParserUtil.FROM;
@@ -144,7 +142,6 @@ import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.StringJoiner;
 import java.util.TreeSet;
 
 import org.h2.api.ErrorCode;
@@ -237,9 +234,7 @@ import org.h2.command.query.TableValueConstructor;
 import org.h2.constraint.ConstraintActionType;
 import org.h2.engine.ConnectionInfo;
 import org.h2.engine.Constants;
-import org.h2.engine.Database;
 import org.h2.engine.DbObject;
-import org.h2.engine.DbSettings;
 import org.h2.engine.IsolationLevel;
 import org.h2.engine.Mode;
 import org.h2.engine.Mode.ModeEnum;
@@ -421,92 +416,21 @@ import org.h2.value.ValueVarchar;
  * @author Noel Grandin
  * @author Nicolas Fortin, Atelier SIG, IRSTV FR CNRS 24888
  */
-public class Parser {
+public final class Parser extends ParserBase {
 
     private static final String WITH_STATEMENT_SUPPORTS_LIMITED_SUB_STATEMENTS =
             "WITH statement supports only SELECT, TABLE, VALUES, " +
             "CREATE TABLE, INSERT, UPDATE, MERGE or DELETE statements";
 
-    private final Database database;
-    private final SessionLocal session;
-
-    /**
-     * @see org.h2.engine.DbSettings#databaseToLower
-     */
-    private final boolean identifiersToLower;
-    /**
-     * @see org.h2.engine.DbSettings#databaseToUpper
-     */
-    private final boolean identifiersToUpper;
-
-    /**
-     * @see org.h2.engine.SessionLocal#isVariableBinary()
-     */
-    private final boolean variableBinary;
-
-    private final BitSet nonKeywords;
-
-    ArrayList<Token> tokens;
-    int tokenIndex;
-    Token token;
-    private int currentTokenType;
-    private String currentToken;
-    private String sqlCommand;
     private CreateView createView;
     private Prepared currentPrepared;
     private Select currentSelect;
     private List<TableView> cteCleanups;
-    private ArrayList<Parameter> parameters;
-    private BitSet usedParameters = new BitSet();
     private String schemaName;
-    private ArrayList<String> expectedList;
     private boolean rightsChecked;
     private boolean recompileAlways;
-    private boolean literalsChecked;
     private int orderInFrom;
     private boolean parseDomainConstraint;
-
-    /**
-     * Parses the specified collection of non-keywords.
-     *
-     * @param nonKeywords array of non-keywords in upper case
-     * @return bit set of non-keywords, or {@code null}
-     */
-    public static BitSet parseNonKeywords(String[] nonKeywords) {
-        if (nonKeywords.length == 0) {
-            return null;
-        }
-        BitSet set = new BitSet();
-        for (String nonKeyword : nonKeywords) {
-            int index = Arrays.binarySearch(TOKENS, FIRST_KEYWORD, LAST_KEYWORD + 1, nonKeyword);
-            if (index >= 0) {
-                set.set(index);
-            }
-        }
-        return set.isEmpty() ? null : set;
-    }
-
-    /**
-     * Formats a comma-separated list of keywords.
-     *
-     * @param nonKeywords bit set of non-keywords, or {@code null}
-     * @return comma-separated list of non-keywords
-     */
-    public static String formatNonKeywords(BitSet nonKeywords) {
-        if (nonKeywords == null || nonKeywords.isEmpty()) {
-            return "";
-        }
-        StringBuilder builder = new StringBuilder();
-        for (int i = -1; (i = nonKeywords.nextSetBit(i + 1)) >= 0;) {
-            if (i >= FIRST_KEYWORD && i <= LAST_KEYWORD) {
-                if (builder.length() > 0) {
-                    builder.append(',');
-                }
-                builder.append(TOKENS[i]);
-            }
-        }
-        return builder.toString();
-    }
 
     /**
      * Creates a new instance of parser.
@@ -514,25 +438,7 @@ public class Parser {
      * @param session the session
      */
     public Parser(SessionLocal session) {
-        this.database = session.getDatabase();
-        DbSettings settings = database.getSettings();
-        this.identifiersToLower = settings.databaseToLower;
-        this.identifiersToUpper = settings.databaseToUpper;
-        this.variableBinary = session.isVariableBinary();
-        this.nonKeywords = session.getNonKeywords();
-        this.session = session;
-    }
-
-    /**
-     * Creates a new instance of parser for special use cases.
-     */
-    public Parser() {
-        database = null;
-        identifiersToLower = false;
-        identifiersToUpper = false;
-        variableBinary = false;
-        nonKeywords = null;
-        session = null;
+        super(session);
     }
 
     /**
@@ -641,17 +547,6 @@ public class Parser {
             command.clearCTE();
             throw t;
         }
-    }
-
-    private ArrayList<Token> getRemainingTokens(int offset) {
-        List<Token> subList = tokens.subList(tokenIndex, tokens.size());
-        ArrayList<Token> remainingTokens = new ArrayList<>(subList);
-        subList.clear();
-        tokens.add(new Token.EndOfInputToken(offset));
-        for (Token token : remainingTokens) {
-            token.subtractFromStart(offset);
-        }
-        return remainingTokens;
     }
 
     /**
@@ -901,13 +796,6 @@ public class Parser {
             setSQL(c, start);
         }
         return c;
-    }
-
-    private DbException getSyntaxError() {
-        if (expectedList == null || expectedList.isEmpty()) {
-            return DbException.getSyntaxError(sqlCommand, token.start());
-        }
-        return DbException.getSyntaxError(sqlCommand, token.start(), String.join(", ", expectedList));
     }
 
     private Prepared parseBackup() {
@@ -1298,19 +1186,6 @@ public class Parser {
             return table.getRowIdColumn();
         }
         return table.getColumn(readIdentifier());
-    }
-
-    /**
-     * Read comma or closing brace.
-     *
-     * @return {@code true} if comma is read, {@code false} if brace is read
-     */
-    private boolean readIfMore() {
-        if (readIf(COMMA)) {
-            return true;
-        }
-        read(CLOSE_PAREN);
-        return false;
     }
 
     private Prepared parseHelp() {
@@ -1848,10 +1723,10 @@ public class Parser {
                 return readCorrelation(tableFilter);
             }
         } else if (readIf(VALUES)) {
-            BitSet outerUsedParameters = initParametersScope();
+            BitSet outerUsedParameters = openParametersScope();
             TableValueConstructor query = parseValues();
             alias = session.getNextSystemIdentifier(sqlCommand);
-            table = query.toTable(alias, null, getUsedParameters(outerUsedParameters), createView != null,
+            table = query.toTable(alias, null, closeParametersScope(outerUsedParameters), createView != null,
                     currentSelect);
         } else if (readIf(TABLE, OPEN_PAREN)) {
             // Table function derived table
@@ -1940,9 +1815,9 @@ public class Parser {
     }
 
     private TableFilter readDerivedTableWithCorrelation() {
-        BitSet outerUsedParameters = initParametersScope();
+        BitSet outerUsedParameters = openParametersScope();
         Query query = parseQueryExpression();
-        ArrayList<Parameter> queryParameters = getUsedParameters(outerUsedParameters);
+        ArrayList<Parameter> queryParameters = closeParametersScope(outerUsedParameters);
         read(CLOSE_PAREN);
         Table table;
         String alias;
@@ -2620,18 +2495,18 @@ public class Parser {
     }
 
     private Query parseQuery() {
-        BitSet outerUsedParameters = initParametersScope();
+        BitSet outerUsedParameters = openParametersScope();
         Query command = parseQueryExpression();
-        ArrayList<Parameter> params = getUsedParameters(outerUsedParameters);
+        ArrayList<Parameter> params = closeParametersScope(outerUsedParameters);
         command.setParameterList(params);
         command.init();
         return command;
     }
 
     private Prepared parseWithStatementOrQuery(int start) {
-        BitSet outerUsedParameters = initParametersScope();
+        BitSet outerUsedParameters = openParametersScope();
         Prepared command = parseWith();
-        ArrayList<Parameter> params = getUsedParameters(outerUsedParameters);
+        ArrayList<Parameter> params = closeParametersScope(outerUsedParameters);
         command.setParameterList(params);
         if (command instanceof Query) {
             Query query = (Query) command;
@@ -2937,7 +2812,7 @@ public class Parser {
         Select command = new Select(session, currentSelect);
         Select oldSelect = currentSelect;
         Prepared oldPrepared = currentPrepared;
-        BitSet outerUsedParameters = initParametersScope();
+        BitSet outerUsedParameters = openParametersScope();
         currentSelect = command;
         currentPrepared = command;
         parseSelectExpressions(command);
@@ -3010,7 +2885,7 @@ public class Parser {
             command.setWindowQuery();
             command.setQualify(readExpressionWithGlobalConditions());
         }
-        command.setParameterList(getUsedParameters(outerUsedParameters));
+        command.setParameterList(closeParametersScope(outerUsedParameters));
         currentSelect = oldSelect;
         currentPrepared = oldPrepared;
         setSQL(command, start);
@@ -4956,33 +4831,6 @@ public class Parser {
         }
     }
 
-    private Parameter readParameter() {
-        int index = ((Token.ParameterToken) token).index() - 1;
-        read();
-        usedParameters.set(index);
-        return parameters.get(index);
-    }
-
-    private BitSet initParametersScope() {
-        BitSet outerUsedParameters = usedParameters;
-        usedParameters = new BitSet();
-        return outerUsedParameters;
-    }
-
-    private ArrayList<Parameter> getUsedParameters(BitSet outerUsedParameters) {
-        BitSet innerUsedParameters = usedParameters;
-        int size = innerUsedParameters.cardinality();
-        ArrayList<Parameter> params = new ArrayList<>(size);
-        if (size > 0) {
-            for (int i = -1; (i = innerUsedParameters.nextSetBit(i + 1)) >= 0;) {
-                params.add(parameters.get(i));
-            }
-        }
-        outerUsedParameters.or(innerUsedParameters);
-        usedParameters = outerUsedParameters;
-        return params;
-    }
-
     private Expression readTerm() {
         Expression r = currentTokenType == IDENTIFIER ? readTermWithIdentifier() : readTermWithoutIdentifier();
         for (;;) {
@@ -5605,88 +5453,6 @@ public class Parser {
         return whenOperand;
     }
 
-    private int readNonNegativeInt() {
-        int v = readInt();
-        if (v < 0) {
-            throw DbException.getInvalidValueException("non-negative integer", v);
-        }
-        return v;
-    }
-
-    private int readInt() {
-        boolean minus = false;
-        if (currentTokenType == MINUS_SIGN) {
-            minus = true;
-            read();
-        } else if (currentTokenType == PLUS_SIGN) {
-            read();
-        }
-        if (currentTokenType != LITERAL) {
-            throw DbException.getSyntaxError(sqlCommand, token.start(), "integer");
-        }
-        Value value = token.value(session);
-        if (minus) {
-            // must do that now, otherwise Integer.MIN_VALUE would not work
-            value = value.negate();
-        }
-        int i = value.getInt();
-        read();
-        return i;
-    }
-
-    private long readPositiveLong() {
-        long v = readLong();
-        if (v <= 0) {
-            throw DbException.getInvalidValueException("positive long", v);
-        }
-        return v;
-    }
-
-    private long readLong() {
-        boolean minus = false;
-        if (currentTokenType == MINUS_SIGN) {
-            minus = true;
-            read();
-        } else if (currentTokenType == PLUS_SIGN) {
-            read();
-        }
-        if (currentTokenType != LITERAL) {
-            throw DbException.getSyntaxError(sqlCommand, token.start(), "long");
-        }
-        Value value = token.value(session);
-        if (minus) {
-            // must do that now, otherwise Long.MIN_VALUE would not work
-            value = value.negate();
-        }
-        long i = value.getLong();
-        read();
-        return i;
-    }
-
-    private boolean readBooleanSetting() {
-        switch (currentTokenType) {
-        case ON:
-        case TRUE:
-            read();
-            return true;
-        case FALSE:
-            read();
-            return false;
-        case LITERAL:
-            boolean result = token.value(session).getBoolean();
-            read();
-            return result;
-        }
-        if (readIf("OFF")) {
-            return false;
-        } else {
-            if (expectedList != null) {
-                addMultipleExpected(ON, TRUE, FALSE);
-            }
-            throw getSyntaxError();
-        }
-    }
-
     private String readString() {
         int sqlIndex = token.start();
         Expression expr = readExpression();
@@ -5781,276 +5547,6 @@ public class Parser {
         String s = currentToken;
         read();
         return s;
-    }
-
-    private void read(String expected) {
-        if (!testToken(expected, token)) {
-            addExpected(expected);
-            throw getSyntaxError();
-        }
-        read();
-    }
-
-    private void read(int tokenType) {
-        if (tokenType != currentTokenType) {
-            addExpected(tokenType);
-            throw getSyntaxError();
-        }
-        read();
-    }
-
-    private boolean readIf(String tokenName) {
-        if (testToken(tokenName, token)) {
-            read();
-            return true;
-        }
-        addExpected(tokenName);
-        return false;
-    }
-
-    private boolean readIf(String tokenName1, String tokenName2) {
-        int i = tokenIndex + 1;
-        if (i + 1 < tokens.size() && testToken(tokenName1, token) && testToken(tokenName2, tokens.get(i))) {
-            setTokenIndex(i + 1);
-            return true;
-        }
-        addExpected(tokenName1, tokenName2);
-        return false;
-    }
-
-    private boolean readIf(String tokenName1, int tokenType2) {
-        int i = tokenIndex + 1;
-        if (i + 1 < tokens.size() && tokens.get(i).tokenType() == tokenType2 && testToken(tokenName1, token)) {
-            setTokenIndex(i + 1);
-            return true;
-        }
-        addExpected(tokenName1, TOKENS[tokenType2]);
-        return false;
-    }
-
-    private boolean readIf(int tokenType) {
-        if (tokenType == currentTokenType) {
-            read();
-            return true;
-        }
-        addExpected(tokenType);
-        return false;
-    }
-
-    private boolean readIf(int tokenType1, int tokenType2) {
-        if (tokenType1 == currentTokenType) {
-            int i = tokenIndex + 1;
-            if (tokens.get(i).tokenType() == tokenType2) {
-                setTokenIndex(i + 1);
-                return true;
-            }
-        }
-        addExpected(tokenType1, tokenType2);
-        return false;
-    }
-
-    private boolean readIf(int tokenType1, String tokenName2) {
-        if (tokenType1 == currentTokenType) {
-            int i = tokenIndex + 1;
-            if (testToken(tokenName2, tokens.get(i))) {
-                setTokenIndex(i + 1);
-                return true;
-            }
-        }
-        addExpected(TOKENS[tokenType1], tokenName2);
-        return false;
-    }
-
-    private boolean readIf(Object... tokensTypesOrNames) {
-        int count = tokensTypesOrNames.length;
-        int size = tokens.size();
-        int i = tokenIndex;
-        check: if (i + count < size) {
-            for (Object tokenTypeOrName : tokensTypesOrNames) {
-                if (!testToken(tokenTypeOrName, tokens.get(i++))) {
-                    break check;
-                }
-            }
-            setTokenIndex(i);
-            return true;
-        }
-        addExpected(tokensTypesOrNames);
-        return false;
-    }
-
-    private boolean isToken(String tokenName) {
-        if (testToken(tokenName, token)) {
-            return true;
-        }
-        addExpected(tokenName);
-        return false;
-    }
-
-    private boolean testToken(Object expected, Token token) {
-        return expected instanceof Integer ? (int) expected == token.tokenType() : testToken((String) expected, token);
-    }
-
-    private boolean testToken(String tokenName, Token token) {
-        if (!token.isQuoted()) {
-            String s = token.asIdentifier();
-            return identifiersToUpper ? tokenName.equals(s) : tokenName.equalsIgnoreCase(s);
-        }
-        return false;
-    }
-
-    private boolean isToken(int tokenType) {
-        if (tokenType == currentTokenType) {
-            return true;
-        }
-        addExpected(tokenType);
-        return false;
-    }
-
-    private boolean equalsToken(String a, String b) {
-        if (a == null) {
-            return b == null;
-        } else
-            return a.equals(b) || !identifiersToUpper && a.equalsIgnoreCase(b);
-    }
-
-    private boolean isIdentifier() {
-        return currentTokenType == IDENTIFIER || nonKeywords != null && nonKeywords.get(currentTokenType);
-    }
-
-    private void addExpected(String token) {
-        if (expectedList != null) {
-            expectedList.add(token);
-        }
-    }
-
-    private void addExpected(int tokenType) {
-        if (expectedList != null) {
-            expectedList.add(TOKENS[tokenType]);
-        }
-    }
-
-    private void addExpected(int tokenType1, int tokenType2) {
-        if (expectedList != null) {
-            expectedList.add(TOKENS[tokenType1] + ' ' + TOKENS[tokenType2]);
-        }
-    }
-
-    private void addExpected(String tokenType1, String tokenType2) {
-        if (expectedList != null) {
-            expectedList.add(tokenType1 + ' ' + tokenType2);
-        }
-    }
-
-    private void addExpected(Object... tokens) {
-        if (expectedList != null) {
-            StringJoiner j = new StringJoiner(" ");
-            for (Object token : tokens) {
-                j.add(token instanceof Integer ? TOKENS[(int) token] : (String) token);
-            }
-            expectedList.add(j.toString());
-        }
-    }
-
-    private void addMultipleExpected(int ... tokenTypes) {
-        for (int tokenType : tokenTypes) {
-            expectedList.add(TOKENS[tokenType]);
-        }
-    }
-
-    private void read() {
-        if (expectedList != null) {
-            expectedList.clear();
-        }
-        int size = tokens.size();
-        if (tokenIndex + 1 < size) {
-            token = tokens.get(++tokenIndex);
-            currentTokenType = token.tokenType();
-            currentToken = token.asIdentifier();
-            if (currentToken != null && currentToken.length() > Constants.MAX_IDENTIFIER_LENGTH) {
-                throw DbException.get(ErrorCode.NAME_TOO_LONG_2, currentToken.substring(0, 32),
-                        "" + Constants.MAX_IDENTIFIER_LENGTH);
-            } else if (currentTokenType == LITERAL) {
-                checkLiterals();
-            }
-        } else {
-            throw getSyntaxError();
-        }
-    }
-
-    private void checkLiterals() {
-        if (!literalsChecked && session != null && !session.getAllowLiterals()) {
-            int allowed = database.getAllowLiterals();
-            if (allowed == Constants.ALLOW_LITERALS_NONE
-                    || ((token instanceof Token.CharacterStringToken || token instanceof Token.BinaryStringToken)
-                            && allowed != Constants.ALLOW_LITERALS_ALL)) {
-                throw DbException.get(ErrorCode.LITERALS_ARE_NOT_ALLOWED);
-            }
-        }
-    }
-
-    private void initialize(String sql, ArrayList<Token> tokens, boolean stopOnCloseParen) {
-        if (sql == null) {
-            sql = "";
-        }
-        sqlCommand = sql;
-        if (tokens == null) {
-            BitSet usedParameters = new BitSet();
-            this.tokens = new Tokenizer(database, identifiersToUpper, identifiersToLower, nonKeywords)
-                    .tokenize(sql, stopOnCloseParen, usedParameters);
-            if (parameters == null) {
-                int l = usedParameters.length();
-                if (l > Constants.MAX_PARAMETER_INDEX) {
-                    throw DbException.getInvalidValueException("parameter index", l);
-                }
-                if (l > 0) {
-                    parameters = new ArrayList<>(l);
-                    for (int i = 0; i < l; i++) {
-                        /*
-                         * We need to create parameters even when they aren't
-                         * actually used, for example, VALUES ?1, ?3 needs
-                         * parameters ?1, ?2, and ?3.
-                         */
-                        parameters.add(new Parameter(i));
-                    }
-                } else {
-                    parameters = new ArrayList<>();
-                }
-            }
-        } else {
-            this.tokens = tokens;
-        }
-        resetTokenIndex();
-    }
-
-    private void resetTokenIndex() {
-        tokenIndex = -1;
-        token = null;
-        currentTokenType = -1;
-        currentToken = null;
-    }
-
-    void setTokenIndex(int index) {
-        if (index != tokenIndex) {
-            if (expectedList != null) {
-                expectedList.clear();
-            }
-            token = tokens.get(index);
-            tokenIndex = index;
-            currentTokenType = token.tokenType();
-            currentToken = token.asIdentifier();
-        }
-    }
-
-    private static boolean isKeyword(int tokenType) {
-        return tokenType >= FIRST_KEYWORD && tokenType <= LAST_KEYWORD;
-    }
-
-    private boolean isKeyword(String s) {
-        return ParserUtil.isKeyword(s, !identifiersToUpper);
-    }
-
-    private String upperName(String name) {
-        return identifiersToUpper ? name : StringUtils.toUpperEnglish(name);
     }
 
     private Column parseColumnForTable(String columnName, boolean defaultNullable) {
@@ -7563,7 +7059,7 @@ public class Parser {
                 isTemporary, session, cteViewName, schema, columns, database);
         List<Column> columnTemplateList;
         String[] querySQLOutput = new String[1];
-        BitSet outerUsedParameters = initParametersScope();
+        BitSet outerUsedParameters = openParametersScope();
         ArrayList<Parameter> queryParameters;
         try {
             read(AS);
@@ -7576,7 +7072,7 @@ public class Parser {
             columnTemplateList = QueryExpressionTable.createQueryColumnTemplateList(cols, withQuery, querySQLOutput);
 
         } finally {
-            queryParameters = getUsedParameters(outerUsedParameters);
+            queryParameters = closeParametersScope(outerUsedParameters);
             TableView.destroyShadowTableForRecursiveExpression(isTemporary, session, recursiveTable);
         }
 
@@ -9751,33 +9247,8 @@ public class Parser {
         }
     }
 
-    /**
-     * Add double quotes around an identifier if required.
-     *
-     * @param s the identifier
-     * @param sqlFlags formatting flags
-     * @return the quoted identifier
-     */
-    public static String quoteIdentifier(String s, int sqlFlags) {
-        if (s == null) {
-            return "\"\"";
-        }
-        if ((sqlFlags & HasSQL.QUOTE_ONLY_WHEN_REQUIRED) != 0 && ParserUtil.isSimpleIdentifier(s, false, false)) {
-            return s;
-        }
-        return StringUtils.quoteIdentifier(s);
-    }
-
-    public void setLiteralsChecked(boolean literalsChecked) {
-        this.literalsChecked = literalsChecked;
-    }
-
     public void setRightsChecked(boolean rightsChecked) {
         this.rightsChecked = rightsChecked;
-    }
-
-    public void setSuppliedParameters(ArrayList<Parameter> suppliedParameters) {
-        this.parameters = suppliedParameters;
     }
 
     /**
@@ -9821,64 +9292,4 @@ public class Parser {
         return readTableOrView();
     }
 
-    /**
-     * Parses a list of column names or numbers in parentheses.
-     *
-     * @param sql the source SQL
-     * @param offset the initial offset
-     * @return the array of column names ({@code String[]}) or numbers
-     *         ({@code int[]})
-     * @throws DbException on syntax error
-     */
-    public Object parseColumnList(String sql, int offset) {
-        initialize(sql, null, true);
-        for (int i = 0, l = tokens.size(); i < l; i++) {
-            if (tokens.get(i).start() >= offset) {
-                setTokenIndex(i);
-                break;
-            }
-        }
-        read(OPEN_PAREN);
-        if (readIf(CLOSE_PAREN)) {
-            return Utils.EMPTY_INT_ARRAY;
-        }
-        if (isIdentifier()) {
-            ArrayList<String> list = Utils.newSmallArrayList();
-            do {
-                if (!isIdentifier()) {
-                    throw getSyntaxError();
-                }
-                list.add(currentToken);
-                read();
-            } while (readIfMore());
-            return list.toArray(new String[0]);
-        } else if (currentTokenType == LITERAL) {
-            ArrayList<Integer> list = Utils.newSmallArrayList();
-            do {
-                list.add(readInt());
-            } while (readIfMore());
-            int count = list.size();
-            int[] array = new int[count];
-            for (int i = 0; i < count; i++) {
-                array[i] = list.get(i);
-            }
-            return array;
-        } else {
-            throw getSyntaxError();
-        }
-    }
-
-    /**
-     * Returns the last parse index.
-     *
-     * @return the last parse index
-     */
-    public int getLastParseIndex() {
-        return token.start();
-    }
-
-    @Override
-    public String toString() {
-        return StringUtils.addAsterisk(sqlCommand, token.start());
-    }
 }
